pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.0",
    "fastapi>=0.125.0",
    "numpy>=2.3.5",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "python-binance>=1.0.34",
    "python-multipart>=0.0.21",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
